Hands-on projects/solution/API/requirements.txt
pymongo
flask
flask-cors
pyjwt
requests
flask-oauthlib
python-dotenv
python-jose
six